+++ src/product/tool-definitions.js
@@ -49,6 +49,7 @@ import { imageWorkflowToolDefinitions } from './definitions/image-workflow-tools
 import { pdfFileToolDefinitions } from './definitions/pdf-file-tools.js';
 import { pdfWorkflowToolDefinitions } from './definitions/pdf-workflow-tools.js';
 import { pdfImageToolDefinitions } from './definitions/pdf-image-tools.js';
+import { pdfDocumentToolDefinitions } from './definitions/pdf-document-tools.js';
 import { videoFileToolDefinitions } from './definitions/video-file-tools.js';
 import { financeDefinitions } from './definitions/finance.js';
 import { healthDefinitions } from './definitions/health.js';
@@ -1068,6 +1069,7 @@ const toolDefinitions = Object.freeze({
     ...pdfFileToolDefinitions,
     ...pdfWorkflowToolDefinitions,
     ...pdfImageToolDefinitions,
+    ...pdfDocumentToolDefinitions,
     ...videoFileToolDefinitions,
 });